html/rendering/non-replaced-elements/tables/table-bordercolor.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Table bordercolor</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/rendering.html#tables-2">
<link rel="author" title="Intel" href="http://www.intel.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<table id="table" bordercolor="red">
  <thead>
    <tr>
      <th>Month</th>
      <th>Savings</th>
    </tr>
  </thead>
  <tbody>
    <tr>
      <td>January</td>
      <td>$60</td>
    </tr>
    <tr>
      <td>February</td>
      <td>$80</td>
    </tr>
  </tbody>
  <tfoot>
    <tr>
      <td>Sum</td>
      <td>$140</td>
    </tr>
  </tfoot>
</table>

<script>

test(() => {
  let elem = document.getElementById("table");
  let css = window.getComputedStyle(elem, null).getPropertyValue("border-color");
  assert_equals(css, "rgb(255, 0, 0)");
}, `table bordercolor attribute is correct`)

</script>
